Add font weight, size and color/background color to HTML elements using class and ID selectors
All odd elements (1,3,5) used the same pink background and font family, so i created a single class (odd-number) to apply
to all these. Element 3 needed to be a bit bigger than the rest, so I added a second class to it (bigger1) and
made the font on it bigger at 24px. For elements 2 and 4, I created ID selectors, because 1. they had specific and
unique color and size traits, and 2. I figured learning how to use and practicing them would be beneficial.