Add a smaller and distorted image in the same space as the originals
For this lesson I edited only the CSS file. The first image with the woman needed to be made smaller, while keeping the
same image porportions, so I set the img height to auto, while shrinking width to 300px. Meanwhile, the second
image needed to be distorted. So I hardcoded the height to be twice that of the width of 400px.