Add CSS to 3/5 of the text lines
the execrcise asked me to give attributes to the 1st, 4th, and 5th lines, that would give it a yellow background
color, red text, and a larger font size. To do this, I used a descendant combinator in between a type selector
and a class selector. This simplified the amount of code needed, and allowed me to use just one declaration
block in applying it to the specific lines.