Alter CSS selectors to get desired outcome look
The second paragraph was not small enough. I eneded up chaining the .para selector to the existing .small-para selector
which prosuced desired outcome. I turned the confirm-button selector into an ID, and chained it the existing button
selector, which gave enough specificity for the green background color to take effect. To bold the last line, I simply
chained the .text selector to the .child selector, which gave it more specificty