Change CSS file to reflect new padding, margin, and placement
I increased int he total amount of padding in Div one and three, because 12px of space in between the
content and border was called for. Div one also needed to have it's margin increased to 12px
to ensure it was separated from every other Div on the page. Margin of Div 2 was also increased.
Lastly, I kept trying out different values for "margin-left" until Div Three lined up perfectly with
the right side of the page.