Format CSS correctly to prouce desired card image
Padding, margin, text alignment, and font size all needed to be altered to rpoduce the desired image given in the
solution image. The key was to create enough padding on .card to show an 8px line of white padding all around the outside
of the card. from there, I had to alter the paddings of the blue sections (the content of the boxes), as well as their
margins from each other.