Format flex items to correct positioning and spacing
Following the instructions and the desired outcome pic of the git-hub lesson, I changed the formatting of the
header using flexbox. Edits were made to the header selector was altered to be in display: flexbox mode
while justify-content and align-items properties were used to align the flex items correctly. Regarding the logo,
margin and padding adjustments were made to match desired outcome. Finally, the ul selector's default padding,
spacing, and items listing was changed to match outcome.