Space out and properly align the header bar
In the HTML, I created two new div containers so that could be treated as nested flex oxntainers. In the stylesheet, flex porperties
and paddings were adjusted to correctly space out the items in the header. The content needed to be pushed to the left and right ends
of the flex xontainer, as well as be centered and sapced evenly from themselves.