Apply spacing, padding, and alignment to pop up box
I used flexbox to change the flex items to the desired layout. The HTML needed some editing as well,
as the close window icon needed to be nested within the heading class. This would put the two items
into the same "container". Meanwhile, the rest of the items, besides the exclamation point icon, were
also put into a separate container, helping to display the proper alignment and stacking of items.