Spread and properly align links while centering content
Once again flex box was used to get the correct positioning of the flex items. The header was pushed to the top, content pushed
to the middle, and footer pushed to the bottom, by changing the flex direction to column. Links for the right and left side were
gapped and then properly aligned with each side to create an equal, centralized look. The content container was set to
flex-direction: column; as well to get the items to stack on top of each other in the middle of screen